% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionWeights-package.R
\docType{package}
\name{SelectionWeights-package}
\alias{SelectionWeights}
\alias{SelectionWeights-package}
\title{SelectionWeights: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://isglobal-cep.github.io/SelectionWeights/}
}

}
\author{
\strong{Maintainer}: Lorenzo Fabbri \email{lorenzo.fabbri@isglobal.org} (\href{https://orcid.org/0000-0003-3031-322X}{ORCID})

}
\keyword{internal}
